+++ components/ContactCTA.tsx
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import { Button } from "@material-tailwind/react";
 import { motion } from "framer-motion";
