+++ components/HeroIcon.tsx
