+++ components/MinistryCard.tsx
