+++ components/index.ts
@@ -1,0 +1,19 @@
+import Footer from "./Footer";
+import ContactCTA from "./ContactCTA";
+import HeroIcon from "./HeroIcon";
+import MinistryCard from "./MinistryCard";
+import ScrollToTop from "./ScrollToTop";
+import ScrollUp from "./Common/ScrollUp";
+import SectionHeading from "./SectionHeading";
+import NavigationBar from "./NavigationBar";
+
+export {
+  NavigationBar,
+  Footer,
+  ContactCTA,
+  HeroIcon,
+  MinistryCard,
+  ScrollToTop,
+  SectionHeading,
+  ScrollUp,
+};
