--- components/page-wrapper.tsx
@@ -1,19 +1,0 @@
-"use client";
-import { motion, AnimatePresence } from "framer-motion";
-
-const pageWrapper = ({ children }) => (
-  <>
-    <AnimatePresence>
-      <motion.div
-        initial={{ opacity: 0, y: 15 }}
-        animate={{ opacity: 1, y: 0 }}
-        exit={{ opacity: 0, y: 15 }}
-        transition={{ delay: 0.25 }}
-      >
-        {children}
-      </motion.div>
-    </AnimatePresence>
-  </>
-);
-
-export default pageWrapper;
