+++ sections/index.ts
@@ -1,0 +1,19 @@
+import Hero from "./Hero";
+import OurChurch from "./OurChurch";
+import YearlyTheme from "./YearlyTheme";
+import MissionHeading from "./MissionHeading";
+import MissionBody from "./MissionBody";
+import OurLeaders from "./OurLeaders";
+import OurMinistries from "./OurMinistries";
+import Contact from "./Contact";
+
+export {
+  Hero,
+  OurChurch,
+  YearlyTheme,
+  MissionHeading,
+  MissionBody,
+  OurLeaders,
+  OurMinistries,
+  Contact,
+};
